Set up AbstractBip39 case with data providers
